Added Midi Message Pages
This change adds 3 additional pages of midi messages assigned to the 4 faders. These pages are accessed through brief presses of the 'message' and 'channel' buttons, each acting as a previous and next button respectively.

The pages are broken out as follows:
Page 1: 1 2 3 4
Page 2: 5 6 7 8
Page 3: A B C D
Page 4: E F G H

In order to make the added messages easier to differentiate and to allow the user to quickly identify the page they are on, the top-left 7-segment display will show the name of the (i.e., fader 3 while on page 2 would show C).

One other change worth noting is that the midi channel set by the first fader is what gets used by all 16 messages in this code. If this is too rigid, a small note was added to line 520 with quick ways to open this back up.

I'm pretty happy with how this turned out, basically quadrupling the midi output potential. Excited to try this out with things like an Empress Zoia or a 1010 Box.